Revert "fix(ui): edited the calendar view and list view design"
